# ssp-backend-flask/swagger_server/models/create_game.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.player import Player  # noqa: F401,E501
from swagger_server import util


class CreateGame(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, host: Player=None):  # noqa: E501
        """CreateGame - a model defined in Swagger

        :param host: The host of this CreateGame.  # noqa: E501
        :type host: Player
        """
        self.swagger_types = {
            'host': Player
        }

        self.attribute_map = {
            'host': 'host'
        }
        self._host = host

    @classmethod
    def from_dict(cls, dikt) -> 'CreateGame':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateGame of this CreateGame.  # noqa: E501
        :rtype: CreateGame
        """
        return util.deserialize_model(dikt, cls)

    @property
    def host(self) -> Player:
        """Gets the host of this CreateGame.


        :return: The host of this CreateGame.
        :rtype: Player
        """
        return self._host

    @host.setter
    def host(self, host: Player):
        """Sets the host of this CreateGame.


        :param host: The host of this CreateGame.
        :type host: Player
        """
        if host is None:
            raise ValueError("Invalid value for `host`, must not be `None`")  # noqa: E501

        self._host = host