# ssp-backend-flask/swagger_server/models/game_status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class GameStatus(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    """
    allowed enum values
    """
    WAITING_FOR_GUEST = "WAITING_FOR_GUEST"
    SETUP_PLACE_FLAG = "SETUP_PLACE_FLAG"
    SETUP_PLACE_TRAP = "SETUP_PLACE_TRAP"
    SETUP_SHUFFLE_PAWNS = "SETUP_SHUFFLE_PAWNS"
    RUNNING_PLAYER_TURN = "RUNNING_PLAYER_TURN"
    RUNNING_OPPONENT_TURN = "RUNNING_OPPONENT_TURN"
    RUNNING_FIGHT = "RUNNING_FIGHT"
    RUNNING_SAME_WEAPON = "RUNNING_SAME_WEAPON"
    FINISHED_HOST_WIN = "FINISHED_HOST_WIN"
    FINISHED_GUEST_WIN = "FINISHED_GUEST_WIN"
    FINISHED_DRAW = "FINISHED_DRAW"
    def __init__(self):  # noqa: E501
        """GameStatus - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'GameStatus':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The GameStatus of this GameStatus.  # noqa: E501
        :rtype: GameStatus
        """
        return util.deserialize_model(dikt, cls)